training/pyproject.toml
[project]
name = "ResearchTraining"
version = "0.1.0"
requires-python = "==3.11"

dependencies = [
  "torch==2.11.0",
  "torchmetrics[detection]",
  "torchvision",
  "transformers",
  "qwen-vl-utils==0.0.14",
  "ultralytics",

  "python-dotenv",
  "mlflow",
  "tqdm",
  "PyYAML",
  "accelerate",
  "ipykernel",
  "requests[socks]",
  "httpx[socks]"
]

[tool.setuptools.packages.find]
where = ["src"]

